Split arearange series into multiple files
