+++ ts/Series/AreaRange/AreaRangePointOptions.d.ts
@@ -1,0 +1,36 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type AreaPointOptions from '../Area/AreaPointOptions';
+
+/* *
+ *
+ *  Declaration
+ *
+ * */
+
+export interface AreaRangePointOptions extends AreaPointOptions {
+    high?: number;
+    low?: number;
+}
+
+/* *
+ *
+ *  Default export
+ *
+ * */
+
+export default AreaRangePointOptions;
